Move tests from util_test.go to curl_cmd_test.
